Melee crit chance display fix
Melee crit chance wasn't being displayed correctly if specific-hand crit
chances were involved, eg. if a player was wearing an axe in their main
hand and a sword in the off hand and only got a crit chance increase to
the axe. It now displays the largest melee crit value the player has,
which is how it should be.